+++ src/main/java/xyz/nikitacartes/easyauth/mixin/PlayerSaveHandlerMixin.java
@@ -5,6 +5,9 @@ import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.NbtIo;
 import net.minecraft.nbt.NbtSizeTracker;
+//? if >= 1.21.9 {
+import net.minecraft.server.PlayerConfigEntry;
+//?}
 import net.minecraft.util.Uuids;
 import net.minecraft.world.PlayerSaveHandler;
 import org.spongepowered.asm.mixin.Final;
@@ -13,12 +16,9 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
-import xyz.nikitacartes.easyauth.utils.PlayerAuth;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.Locale;
 import java.util.Optional;
 
 import static xyz.nikitacartes.easyauth.EasyAuth.*;
@@ -38,29 +38,40 @@ public class PlayerSaveHandlerMixin {
      * @param mixinFile
      */
     @Inject(
-            method = "loadPlayerData(Lnet/minecraft/entity/player/PlayerEntity;Ljava/lang/String;)Ljava/util/Optional;",
+            //? if >= 1.21.9 {
+            method = "loadPlayerData(Lnet/minecraft/server/PlayerConfigEntry;Ljava/lang/String;)Ljava/util/Optional;",
+            //?} else {
+            /*method = "loadPlayerData(Lnet/minecraft/entity/player/PlayerEntity;Ljava/lang/String;)Ljava/util/Optional;",
+            *///?}
             at = @At(
                     value = "INVOKE",
                     target = "Ljava/io/File;exists()Z"
             ),
             cancellable = true
     )
-    private void fileExists(PlayerEntity player, String extension, CallbackInfoReturnable<Optional<NbtCompound>> cir, @Local File mixinFile) {
+    //? if >= 1.21.9 {
+    private void fileExists(PlayerConfigEntry playerConfigEntry, String extension, CallbackInfoReturnable<Optional<NbtCompound>> cir, @Local File mixinFile) {
+    //?} else {
+    /*private void fileExists(PlayerEntity player, String extension, CallbackInfoReturnable<Optional<NbtCompound>> cir, @Local File mixinFile) {
+    *///?}
         if (!(mixinFile.exists() && mixinFile.isFile())) {
-            String playername = player.getGameProfile().getName().toLowerCase(Locale.ENGLISH);
-            PlayerAuth playerAuth = (PlayerAuth) player;
-            if (Boolean.parseBoolean(serverProp.getProperty("online-mode")) && playerAuth.easyAuth$isUsingMojangAccount()) {
-                LogDebug(String.format("Migrating data for %s", playername));
-                File file = new File(this.playerDataDir, Uuids.getOfflinePlayerUuid(player.getGameProfile().getName()) + extension);
+            //? if >= 1.21.9 {
+            String playerName = playerConfigEntry.name();
+            //?} else {
+            /*String playerName = player.getGameProfile().getName();
+            *///?}
+            if (Boolean.parseBoolean(serverProp.getProperty("online-mode"))) {
+                LogDebug(String.format("Migrating data for %s", playerName));
+                File file = new File(this.playerDataDir, Uuids.getOfflinePlayerUuid(playerName) + extension);
                 if (file.exists() && file.isFile()) try {
                     cir.setReturnValue(Optional.of(NbtIo.readCompressed(file.toPath(), NbtSizeTracker.ofUnlimitedBytes())));
                 } catch (IOException e) {
-                    LogWarn(String.format("Failed to load player data for: %s", playername));
+                    LogWarn(String.format("Failed to load player data for: %s", playerName));
                 }
             } else {
                 LogDebug(
-                        String.format("Not migrating %s, as premium status is '%s' and data file is %s present.",
-                                playername, playerAuth.easyAuth$isUsingMojangAccount(), mixinFile.exists() && mixinFile.isFile() ? "" : "not")
+                        String.format("Not migrating %s, data file is %s present.",
+                                playerName, mixinFile.exists() && mixinFile.isFile() ? "" : "not")
                 );
             }
         }