+++ src/main/java/xyz/nikitacartes/easyauth/mixin/ServerPlayerEntityMixin.java
@@ -81,7 +81,11 @@ public abstract class ServerPlayerEntityMixin extends EntityMixin implements Pla
         if (lastLocation == null) {
             lastLocation = new LastLocation();
         }
-        lastLocation.position = player.getPos();
+        //? if >= 1.21.9 {
+        lastLocation.position = player.getEntityPos();
+        //?} else {
+        /*lastLocation.position = player.getPos();
+        *///?}
         lastLocation.yaw = player.getYaw();
         lastLocation.pitch = player.getPitch();
 
@@ -108,8 +112,8 @@ public abstract class ServerPlayerEntityMixin extends EntityMixin implements Pla
             return;
         }
         if (wasDead) {
-            player.kill(player.getWorld());
-            player.getScoreboard().forEachScore(ScoreboardCriterion.DEATH_COUNT, player, (score) -> score.setScore(score.getScore() - 1));
+            player.kill(player.getEntityWorld());
+            player.getEntityWorld().getScoreboard().forEachScore(ScoreboardCriterion.DEATH_COUNT, player, (score) -> score.setScore(score.getScore() - 1));
             return;
         }
         // Puts player to last saved position
@@ -137,7 +141,11 @@ public abstract class ServerPlayerEntityMixin extends EntityMixin implements Pla
             if (world == null) return;
             Entity entity = world.getEntity(ridingEntityUUID);
             if (entity != null) {
-                player.startRiding(entity, true);
+                //? if >= 1.21.9 {
+                player.startRiding(entity, true, false);
+                //?} else {
+                /*player.startRiding(entity, true);
+                *///?}
             } else {
                 LogDebug("Could not find vehicle for player " + username);
             }
@@ -222,7 +230,6 @@ public abstract class ServerPlayerEntityMixin extends EntityMixin implements Pla
         if (authenticated) {
             kickTimer = config.kickTimeout * 20;
             // Updating blocks if needed (in case if portal rescue action happened)
-            World world = player.getWorld();
             BlockPos pos = player.getBlockPos();
 
             // Sending updates to portal blocks